requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.25.0
faiss-cpu==1.7.4
pinecone-client==2.2.2
pytesseract==0.3.10
poppler-utils==0.1.0
pdfplumber==0.10.2
python-docx==0.8.11
Pillow==10.0.0
openpyxl==3.1.2
pytz==2023.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
langchain==0.0.267
openai==0.27.8
transformers==4.32.1
torch==2.0.1
numpy==1.25.2
pandas==2.1.0
scipy==1.11.2
sentencepiece==0.1.99
tenacity==8.2.3
requests==2.31.0
aiohttp==3.8.5
asyncio==3.4.3
json5==0.9.14
email-validator==2.0.0.post2